refactor(multipooler): move rpc functions into it's own file
