+++ go/multipooler/manager/manager_test.go
@@ -29,7 +29,6 @@ import (
 	"github.com/multigres/multigres/go/clustermetadata/topo/memorytopo"
 	"github.com/multigres/multigres/go/mterrors"
 	"github.com/multigres/multigres/go/servenv"
-	"github.com/multigres/multigres/go/test/utils"
 
 	clustermetadatapb "github.com/multigres/multigres/go/pb/clustermetadata"
 	mtrpcpb "github.com/multigres/multigres/go/pb/mtrpc"
@@ -408,89 +407,6 @@ func TestValidateAndUpdateTerm(t *testing.T) {
 	}
 }
 
-func TestPrimaryPosition(t *testing.T) {
-	ctx := context.Background()
-	logger := slog.New(slog.NewTextHandler(os.Stdout, nil))
-
-	serviceID := &clustermetadatapb.ID{
-		Component: clustermetadatapb.ID_MULTIPOOLER,
-		Cell:      "zone1",
-		Name:      "test-service",
-	}
-
-	tests := []struct {
-		name          string
-		poolerType    clustermetadatapb.PoolerType
-		expectError   bool
-		expectedCode  mtrpcpb.Code
-		errorContains string
-	}{
-		{
-			name:          "REPLICA pooler returns FAILED_PRECONDITION",
-			poolerType:    clustermetadatapb.PoolerType_REPLICA,
-			expectError:   true,
-			expectedCode:  mtrpcpb.Code_FAILED_PRECONDITION,
-			errorContains: "pooler type is REPLICA",
-		},
-		{
-			name:          "PRIMARY pooler passes type check",
-			poolerType:    clustermetadatapb.PoolerType_PRIMARY,
-			expectError:   true,
-			errorContains: "database connection failed", // Will fail on DB connection, not type check
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ts, _ := memorytopo.NewServerAndFactory(ctx, "zone1")
-			defer ts.Close()
-
-			// Create temp directory for pooler-dir
-			poolerDir := t.TempDir()
-
-			multipooler := &clustermetadatapb.MultiPooler{
-				Id:            serviceID,
-				Database:      "testdb",
-				Hostname:      "localhost",
-				PortMap:       map[string]int32{"grpc": 8080},
-				Type:          tt.poolerType,
-				ServingStatus: clustermetadatapb.PoolerServingStatus_SERVING,
-			}
-			require.NoError(t, ts.CreateMultiPooler(ctx, multipooler))
-
-			config := &Config{
-				TopoClient: ts,
-				ServiceID:  serviceID,
-				PoolerDir:  poolerDir,
-			}
-			manager := NewMultiPoolerManager(logger, config)
-			defer manager.Close()
-
-			// Start and wait for ready
-			senv := servenv.NewServEnv()
-			go manager.Start(senv)
-			require.Eventually(t, func() bool {
-				return manager.GetState() == ManagerStateReady
-			}, 5*time.Second, 100*time.Millisecond, "Manager should reach Ready state")
-
-			// Call PrimaryPosition
-			_, err := manager.PrimaryPosition(ctx)
-
-			if tt.expectError {
-				require.Error(t, err)
-				assert.Contains(t, err.Error(), tt.errorContains)
-
-				if tt.expectedCode != 0 {
-					code := mterrors.Code(err)
-					assert.Equal(t, tt.expectedCode, code)
-				}
-			} else {
-				require.NoError(t, err)
-			}
-		})
-	}
-}
-
 func TestValidateSyncReplicationParams(t *testing.T) {
 	tests := []struct {
 		name        string
@@ -619,196 +535,6 @@ func TestValidateSyncReplicationParams(t *testing.T) {
 			} else {
 				require.NoError(t, err)
 			}
-		})
-	}
-}
-
-func TestActionLock_MutationMethodsTimeout(t *testing.T) {
-	ctx := context.Background()
-	logger := slog.New(slog.NewTextHandler(os.Stdout, nil))
-
-	serviceID := &clustermetadatapb.ID{
-		Component: clustermetadatapb.ID_MULTIPOOLER,
-		Cell:      "zone1",
-		Name:      "test-service",
-	}
-
-	ts, _ := memorytopo.NewServerAndFactory(ctx, "zone1")
-	defer ts.Close()
-
-	poolerDir := t.TempDir()
-
-	// Create PRIMARY multipooler for testing
-	multipooler := &clustermetadatapb.MultiPooler{
-		Id:            serviceID,
-		Database:      "testdb",
-		Hostname:      "localhost",
-		PortMap:       map[string]int32{"grpc": 8080},
-		Type:          clustermetadatapb.PoolerType_PRIMARY,
-		ServingStatus: clustermetadatapb.PoolerServingStatus_SERVING,
-	}
-	require.NoError(t, ts.CreateMultiPooler(ctx, multipooler))
-
-	config := &Config{
-		TopoClient: ts,
-		ServiceID:  serviceID,
-		PoolerDir:  poolerDir,
-	}
-	manager := NewMultiPoolerManager(logger, config)
-	defer manager.Close()
-
-	// Start and wait for ready
-	senv := servenv.NewServEnv()
-	go manager.Start(senv)
-	require.Eventually(t, func() bool {
-		return manager.GetState() == ManagerStateReady
-	}, 5*time.Second, 100*time.Millisecond, "Manager should reach Ready state")
-
-	// Helper function to hold the lock for a duration
-	holdLock := func(duration time.Duration) context.CancelFunc {
-		lockCtx, cancel := context.WithDeadline(context.Background(), time.Now().Add(duration))
-		lockAcquired := make(chan struct{})
-		go func() {
-			if err := manager.lock(lockCtx); err == nil {
-				// Signal that the lock was acquired
-				close(lockAcquired)
-				// Hold the lock for the duration or until cancelled
-				<-lockCtx.Done()
-				manager.unlock()
-			}
-		}()
-		// Wait for the lock to be acquired
-		<-lockAcquired
-		return cancel
-	}
-
-	tests := []struct {
-		name       string
-		poolerType clustermetadatapb.PoolerType
-		callMethod func(context.Context) error
-	}{
-		{
-			name:       "SetPrimaryConnInfo times out when lock is held",
-			poolerType: clustermetadatapb.PoolerType_REPLICA,
-			callMethod: func(ctx context.Context) error {
-				return manager.SetPrimaryConnInfo(ctx, "localhost", 5432, false, false, 1, true)
-			},
-		},
-		{
-			name:       "StartReplication times out when lock is held",
-			poolerType: clustermetadatapb.PoolerType_REPLICA,
-			callMethod: func(ctx context.Context) error {
-				return manager.StartReplication(ctx)
-			},
-		},
-		{
-			name:       "StopReplication times out when lock is held",
-			poolerType: clustermetadatapb.PoolerType_REPLICA,
-			callMethod: func(ctx context.Context) error {
-				return manager.StopReplication(ctx)
-			},
-		},
-		{
-			name:       "ResetReplication times out when lock is held",
-			poolerType: clustermetadatapb.PoolerType_REPLICA,
-			callMethod: func(ctx context.Context) error {
-				return manager.ResetReplication(ctx)
-			},
-		},
-		{
-			name:       "ConfigureSynchronousReplication times out when lock is held",
-			poolerType: clustermetadatapb.PoolerType_PRIMARY,
-			callMethod: func(ctx context.Context) error {
-				return manager.ConfigureSynchronousReplication(
-					ctx,
-					multipoolermanagerdatapb.SynchronousCommitLevel_SYNCHRONOUS_COMMIT_ON,
-					multipoolermanagerdatapb.SynchronousMethod_SYNCHRONOUS_METHOD_FIRST,
-					1,
-					[]*clustermetadatapb.ID{serviceID},
-					true,
-				)
-			},
-		},
-		{
-			name:       "SetReadOnly times out when lock is held",
-			poolerType: clustermetadatapb.PoolerType_PRIMARY,
-			callMethod: func(ctx context.Context) error {
-				return manager.SetReadOnly(ctx)
-			},
-		},
-		{
-			name:       "ChangeType times out when lock is held",
-			poolerType: clustermetadatapb.PoolerType_PRIMARY,
-			callMethod: func(ctx context.Context) error {
-				return manager.ChangeType(ctx, "REPLICA")
-			},
-		},
-		{
-			name:       "Demote times out when lock is held",
-			poolerType: clustermetadatapb.PoolerType_PRIMARY,
-			callMethod: func(ctx context.Context) error {
-				_, err := manager.Demote(ctx, 1, 5*time.Second, false)
-				return err
-			},
-		},
-		{
-			name:       "UndoDemote times out when lock is held",
-			poolerType: clustermetadatapb.PoolerType_PRIMARY,
-			callMethod: func(ctx context.Context) error {
-				return manager.UndoDemote(ctx)
-			},
-		},
-		{
-			name:       "Promote times out when lock is held",
-			poolerType: clustermetadatapb.PoolerType_REPLICA,
-			callMethod: func(ctx context.Context) error {
-				_, err := manager.Promote(ctx, 1, "", nil, false)
-				return err
-			},
-		},
-		{
-			name:       "SetTerm times out when lock is held",
-			poolerType: clustermetadatapb.PoolerType_PRIMARY,
-			callMethod: func(ctx context.Context) error {
-				return manager.SetTerm(ctx, &multipoolermanagerdatapb.ConsensusTerm{CurrentTerm: 5})
-			},
-		},
-		{
-			name:       "UpdateSynchronousStandbyList times out when lock is held",
-			poolerType: clustermetadatapb.PoolerType_PRIMARY,
-			callMethod: func(ctx context.Context) error {
-				return manager.UpdateSynchronousStandbyList(ctx, multipoolermanagerdatapb.StandbyUpdateOperation_STANDBY_UPDATE_OPERATION_ADD, []*clustermetadatapb.ID{serviceID}, true, 0, true)
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Update the pooler type if needed for this test
-			if tt.poolerType != multipooler.Type {
-				updatedMultipooler, err := ts.UpdateMultiPoolerFields(ctx, serviceID, func(mp *clustermetadatapb.MultiPooler) error {
-					mp.Type = tt.poolerType
-					return nil
-				})
-				require.NoError(t, err)
-				manager.mu.Lock()
-				manager.multipooler.MultiPooler = updatedMultipooler
-				manager.mu.Unlock()
-			}
-
-			// Hold the lock for 2 seconds
-			cancel := holdLock(2 * time.Second)
-			defer cancel()
-
-			// Try to call the method - it should timeout because lock is held
-			err := tt.callMethod(utils.WithTimeout(t, 500*time.Millisecond))
-
-			// Verify the error is a timeout/context error
-			require.Error(t, err, "Method should fail when lock is held")
-			assert.Contains(t, err.Error(), "failed to acquire action lock", "Error should mention lock acquisition failure")
-
-			// Verify the underlying error is context deadline exceeded
-			assert.ErrorIs(t, err, context.DeadlineExceeded, "Should be a deadline exceeded error")
 		})
 	}
 }
